BSSD {BSSD}	R Documentation

Baysian smoothing spine with dependency

Description

This is the main function for the Bayesian Smoothing Spline with Dependency (BSSD) method. This function generates the posterior samples of trend estimater and other parameters.

Usage

BSSD(y,Ty=12,n.s=10^4,c=50,rho.p=1,e2b=1e-10,r=1,k=2,const0=1,uv=FALSE)

Arguments

y
a vector, the observed value (data) at each equally spaced knot

Ty
period for the seasonal effect of data y, default is 12 for monthly series

n.s
number of posterior samples, default is 10,000

c
hyper-parameter for the Pareto prior of eta

rho.p
choice of prior for rho, see rho.logprior

e2b
small value used for open intervals

r
parameter for ratio-of-uniform sampling

k
parameter for ratio-of-uniform sampling

const0
constant that make sure the the posterior density is not too large to compute in R

uv
logic, default is FALSE. If TRUE, the direct results of ratio-of-uniform sampling are reported.

Details

The posterior samples are drawn with ratio-of-uniform method

Value

data
the data vector used

z.mc
a matrix of posterior samples of trend estimates, ncol=n.s

d0.mc
a vector of posterior samples of varience parameter delta_0

eta.mc
a vector of posterior samples of smoothing parameter eta

rho.mc
a vector of posterior samples of correlation parameter rho

rate.accept
the acceptence rate of ratio-of-uniform sampling

Author(s)

Benqian Zhang

See Also

unemployment, SeaDataSimu

Examples

#Ex1: U.S. Unemployment level Jan 2004 - Dec 2006
data(unemployment)
index=1:36
y=unemployment$y[index]
timelab=unemployment$label[index]
mc.result=BSSD(y=y)
#Trend estimation
z.est = apply(mc.result$z.mc, 1, mean)
#Plot
plot(index,y,type = 'l', lty=3,xaxt="n",ylab = '',xlab="Time (by 2 months)",ylim = c(min(y),max(y)),las=1)
lines(index, z.est, type = "l",lty=1,col='red',xlab='')
at=seq(from = 1, to = length(y), by =2)
axis(1, at, labels=format(timelab[at],"%Y-%m"))
#
#Ex2: Simulated data
t=1:36
Trend=9000*exp(-(t-20)^2/2/10^2)+6000
para=c(12,200,100)
y=SeaDataSimu(n.simu=1,Trend,para,RandomS=FALSE)
mc.result=BSSD(y=y)
#Trend estimation
z.est = apply(mc.result$z.mc, 1, mean)
#Plot
plot(index,y,type = 'l', lty=3,ylab = '',xlab="Time",ylim = c(min(y),max(y)),las=1)
lines(index, z.est, type = "l",lty=1,col='red',xlab='')

[Package BSSD version 0.2.1 Index]