SeaDataSimu {BSSD}	R Documentation

Seasonal data simulator

Description

Simulate data with seasonal effects from sine functions and given trend.

Usage

SeaDataSimu(n.simu=1,TrendSimu,para,RandomS=FALSE)

Arguments

n.simu
Number of series to be simulated

TrendSimu
A vector of given trend

para
A vector of parameters. para=(Ty,s1,s2). Ty is the period of seasonal effect. s1, s2 are variation parameters.

RandomS
Logic. If RandomS = FALSE, all n.simu series with share the same seasonal effect. If RandomS = TRUE, the seasonal pattern will be different for each series.

Details

y_t = m_t + seasonal_t + e_t, e_t ~ N(0,s2^2).

m_t is given by TrendSimu.

seasonal_t = sum_i=1^K A_i*sin(2*pi*i*t/Ty+mu_i), where A_i ~ N(0,s1^2), mu_i ~ unif(0,2*pi), K=[Ty/2].

Value

A vector or matrix of simulated data. For matrix, the number of rows nrow = n.simu.

Examples

DataSim=SeaDataSimu(n.simu=1,rep(0,24),c(6,200,100),RandomS=FALSE)

[Package BSSD version 0.2.1 Index]